pyproject.toml
[project]
name = "copybook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]
